[12300] Update money handling to uint64 Update mail code, send mail need test

Conflicts:
	sql/characters.sql
	sql/mangos.sql
	src/game/Guild.h
	src/game/ItemHandler.cpp
	src/game/Level1.cpp
	src/game/Level2.cpp
	src/game/LootHandler.cpp
	src/game/MailHandler.cpp
	src/game/Player.cpp
	src/game/Player.h
	src/game/SpellEffects.cpp
	src/game/TradeHandler.cpp
	src/game/Unit.cpp
	src/shared/revision_nr.h

+++ src/game/LootHandler.cpp
@@ -264,7 +264,7 @@ void WorldSession::HandleLootMoneyOpcode( WorldPacket & /*recv_data*/ )
                     playersNear.push_back(playerGroup);
             }
 
-            uint32 money_per_player = uint32((pLoot->gold)/(playersNear.size()));
+            uint64 money_per_player = uint32((pLoot->gold) / (playersNear.size()));
 
             for (std::vector<Player*>::const_iterator i = playersNear.begin(); i != playersNear.end(); ++i)
             {
